Ajout d'une fonction permettant de dessiner un objet à huit sommets addOcto(...) fonctionne de la même manière que addTriangle et addQuad (précédement ajoutée).
Dessin d'un pavé à la place de la maison, jardin etc lorsqu'il n'y a pas
de split().

+++ segment.cpp
@@ -2,11 +2,11 @@
 
 Segment::Segment(Vertex u, Vertex v): u(u), v(v) {}
 
-int Segment::length() {
+float Segment::length() {
 	return (u-v).norm();
 }
 
-Segment Segment::reduce(int value) {
+Segment Segment::reduce(float value) {
     float reduc = (float)length()/(float)value;
     return Segment(u,u+((v - u) / reduc));
 }
